Complete project cleanup of current features
