+++ lib/src/sms/services/twilio_sms_service.dart
@@ -1,4 +1,5 @@
 import '../../shared/dto/twilio_creds.dart';
+import '../dto/message.dart';
 import '../dto/sent_sms_data.dart';
 
 abstract class TwilioSMSService {
@@ -8,8 +9,8 @@ abstract class TwilioSMSService {
       required TwilioCreds twilioCreds});
 
   Future<SentSmsData> getSmsList(
-      {required String pageSize, required TwilioCreds? twilioCreds});
+      {required String? pageSize, required TwilioCreds twilioCreds});
 
   Future<Message> getSmsData(
-      {required String messageSID, required TwilioCreds? twilioCreds});
+      {required String messageSID, required TwilioCreds twilioCreds});
 }
